mise en page/css, beaucoup de visuel creation lien de redirection dans article et modify.php inscription redirige directement sur le flux plus coherent

+++ Controleurs/ControleurUser.php
@@ -3,10 +3,10 @@
 final class ControleurUser
 {
     public function defautAction() {
-		
+
 		session_start();
 		if(isset($_SESSION['admin']) && $_SESSION['admin']) {
-			
+
 			$user =  new User();
         	Vue::montrer('user/voir', array('users' =>  $user->read()));
 		}
@@ -15,12 +15,19 @@ final class ControleurUser
 		}
     }
 
-    public function deleteAction() {
+   public function deleteAction() {
 
-      $user =  new User();
-      $user->delete($_GET['id']);
+      if(isset($_SESSION['admin']) && $_SESSION['admin']) {
+		  
+        $user =  new User();
+        $user->delete($_GET['id']);
 
-      Vue::montrer('user/voir', array('users' =>  $user->read()));
+        header('Location: ./user');
+      }
+
+      else {
+        header('Location: ./');
+      }
 
     }
 
@@ -35,11 +42,33 @@ final class ControleurUser
 
     public function updateAction() {
 
-      $user =  new User();
       session_start();
-      $user->update($_POST['username'], $_POST['email'], $_SESSION['id']);
 
-      header('Location: ./flux');
+      $user = new User();
+      $id = $user->loginSuccessById($_SESSION['id'], sha1($_POST['mdp']));
+
+
+      if($id > -1) {
+
+        if(isset($_POST['new_mdp']) and !empty($_POST['new_mdp'])) {
+          $user->update($_POST['username'], $_POST['email'], sha1($_POST['new_mdp']), $_SESSION['id']);
+        }
+        else {
+
+          $user->update($_POST['username'], $_POST['email'], sha1($_POST['mdp']), $_SESSION['id']);
+        }
+
+        header('Location: ./flux');
+
+      }
+      else {
+        $_SESSION['erreur'] = "Mot de passe invalide";
+        header('Location: index.php?ctrl=user&action=modify');
+      }
+
+
+
+
 
     }
 
@@ -56,6 +85,49 @@ final class ControleurUser
       else {
         header('Location: /login');
       }
+
+    }
+
+
+    public function changePwdAction() {
+
+      Vue::montrer('user/password', array());
+
+    }
+
+    public function mailAction() {
+
+      $user = new User();
+      session_start();
+
+      if(isset($_POST['username']) && isset($_POST['email']) && !empty($_POST['username']) && !empty($_POST['email'])) {
+
+        $id = $user->exist($_POST['username'], $_POST['email']);
+        if($id > -1) {
+
+          $newMdp = bin2hex(random_bytes(5));
+
+          $res = mail ($_POST['email'] , "Nouveau identifiant Vanestarre" , "Votre mdp est = ".$newMdp);
+
+          if($res) {
+            $user->setNewMdp(sha1($newMdp), $id);
+          }
+
+          header('Location: ./');
+
+        }
+        else {
+
+          $_SESSION['erreur'] = "Un des champs n'est pas correcte";
+          header('Location: ./index.php?ctrl=user&action=changepwd');
+        }
+      }
+      else {
+
+        $_SESSION['erreur'] = "Remplissez les 2 champs";
+        header('Location: ./index.php?ctrl=user&action=changepwd');
+      }
+
 
     }
